sim_env_requirements.txt
absl-py==2.3.1
aiodns==3.1.1
aiofiles==23.2.1
aiohappyeyeballs==2.4.4
aiohttp==3.10.5
aiohttp-cors==0.8.1
aioitertools==0.11.0
aiosignal==1.3.2
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.11.0
anytree==2.13.0
argcomplete==3.6.3
asteval==1.0.6
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.1.0
awscrt==0.23.4
azure-core==1.28.0
azure-identity==1.13.0
azure-storage-blob==12.17.0
boto3==1.36.1
botocore==1.36.1
casadi==3.7.2
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.1.1
cmeel==0.57.3
cmeel-assimp==5.4.3.1
cmeel-boost==1.83.0
cmeel-console-bridge==1.0.2.3
cmeel-octomap==1.10.0
cmeel-qhull==8.0.2.1
cmeel-tinyxml==2.6.2.3
cmeel-urdfdom==3.1.1.1
cmeel-zlib==1.3.1
coloredlogs==15.0.1
comm==0.2.3
construct==2.10.68
contourpy==1.3.1
coverage==7.4.4
cryptography==44.0.0
cycler==0.11.0
cyclonedds==0.10.2
decorator==5.2.1
egl_probe==1.0.2
eigenpy==3.5.1
einops==0.8.1
evdev==1.9.2
executing==2.2.1
expandvars==1.1.2
Farama-Notifications==0.0.4
fastapi==0.115.7
filelock==3.13.1
flaky==3.8.1
flatbuffers==25.9.23
flatdict==4.0.1
fonttools==4.55.3
frozenlist==1.5.0
fsspec==2024.6.1
gitdb==4.0.12
GitPython==3.1.45
grpcio==1.76.0
gunicorn==22.0.0
gym==0.23.1
gym-notices==0.1.0
gymnasium==1.2.0
h11==0.16.0
h5py==3.15.1
hf-xet==1.2.0
hidapi==0.14.0.post2
hpp-fcl==2.4.4
httptools==0.6.1
huggingface-hub==0.36.0
humanfriendly==10.0
hydra-core==1.3.2
idna==3.10
idna-ssl==1.1.0
imageio==2.37.0
imageio-ffmpeg==0.6.0
iniconfig==2.3.0
ipython==9.6.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.5
-e git+https://github.com/isaac-sim/IsaacLab.git@46dff135f44683f031edf346e544fcfd8456b2bb#egg=isaaclab&subdirectory=source/isaaclab
-e git+https://github.com/isaac-sim/IsaacLab.git@46dff135f44683f031edf346e544fcfd8456b2bb#egg=isaaclab_assets&subdirectory=source/isaaclab_assets
-e git+https://github.com/isaac-sim/IsaacLab.git@46dff135f44683f031edf346e544fcfd8456b2bb#egg=isaaclab_mimic&subdirectory=source/isaaclab_mimic
-e git+https://github.com/isaac-sim/IsaacLab.git@46dff135f44683f031edf346e544fcfd8456b2bb#egg=isaaclab_rl&subdirectory=source/isaaclab_rl
-e git+https://github.com/isaac-sim/IsaacLab.git@46dff135f44683f031edf346e544fcfd8456b2bb#egg=isaaclab_tasks&subdirectory=source/isaaclab_tasks
isaacsim==5.0.0.0
isaacsim-app==5.0.0.0
isaacsim-asset==5.0.0.0
isaacsim-benchmark==5.0.0.0
isaacsim-code-editor==5.0.0.0
isaacsim-core==5.0.0.0
isaacsim-cortex==5.0.0.0
isaacsim-example==5.0.0.0
isaacsim-extscache-kit==5.0.0.0
isaacsim-extscache-kit-sdk==5.0.0.0
isaacsim-extscache-physics==5.0.0.0
isaacsim-gui==5.0.0.0
isaacsim-kernel==5.0.0.0
isaacsim-replicator==5.0.0.0
isaacsim-rl==5.0.0.0
isaacsim-robot==5.0.0.0
isaacsim-robot-motion==5.0.0.0
isaacsim-robot-setup==5.0.0.0
isaacsim-ros1==5.0.0.0
isaacsim-ros2==5.0.0.0
isaacsim-sensor==5.0.0.0
isaacsim-storage==5.0.0.0
isaacsim-template==5.0.0.0
isaacsim-test==5.0.0.0
isaacsim-utils==5.0.0.0
isodate==0.6.1
jedi==0.19.2
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
junitparser==4.0.2
jupyterlab_widgets==3.0.15
killport==1.2.0
kiwisolver==1.4.4
llvmlite==0.42.0
logging-mp==0.1.5
loop-rate-limiters==1.2.0
lxml==6.0.2
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==2.1.3
matplotlib==3.10.0
matplotlib-inline==0.2.1
mdurl==0.1.2
meshcat==0.3.2
ml_dtypes==0.5.3
moviepy==2.2.1
mpmath==1.3.0
msal==1.27.0
msal-extensions==1.0.0
msgpack==1.1.2
multidict==6.1.0
nest-asyncio==1.5.6
networkx==3.3
nlopt==2.7.1
numba==0.59.1
numpy==1.26.0
numpy-quaternion==2023.0.3
nvidia-cublas==13.0.0.19
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti==13.0.48
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc==13.0.48
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime==13.0.48
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cudnn-cu13==9.13.0.50
nvidia-cufft==12.0.0.15
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile==1.15.0.42
nvidia-cufile-cu12==1.11.1.6
nvidia-curand==10.4.0.35
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver==12.0.3.29
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse==12.6.2.49
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-cusparselt-cu13==0.8.0
nvidia-nccl-cu12==2.26.2
nvidia-nccl-cu13==2.27.7
nvidia-nvjitlink==13.0.39
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvshmem-cu13==3.3.24
nvidia-nvtx==13.0.39
nvidia-nvtx-cu12==12.6.77
nvsmi==0.4.2
oauthlib==3.2.2
omegaconf==2.3.0
omniverse-kit==107.3.1.206797
onnx==1.19.1
onnxruntime==1.22.1
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
osqp==0.6.7.post3
outcome==1.3.0.post0
packaging==23.0
pandas==2.3.3
params_proto==2.13.2
parso==0.8.5
pexpect==4.9.0
pillow==11.2.1
pin==2.7.0
pin-pink==3.1.0
Pint==0.20.1
platformdirs==4.5.0
plotly==5.3.1
pluggy==1.6.0
portalocker==2.7.0
prettytable==3.3.0
proglog==0.1.12
prompt_toolkit==3.0.52
propcache==0.2.1
protobuf==6.33.0
psutil==5.9.8
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==22.0.0
pyassimp==5.2.5
pycares==4.3.0
pycparser==2.23
pydantic==2.11.10
pydantic_core==2.33.2
pyglet==1.5.31
Pygments==2.19.2
PyJWT==2.10.1
pyngrok==7.4.1
pynput==1.8.1
pyparsing==3.0.9
pyperclip==1.8.0
pypng==0.20220715.0
PySocks==1.7.1
pytest==8.4.2
pytest-mock==3.15.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.20
python-xlib==0.33
pytransform3d==3.14.3
pytz==2024.1
PyYAML==6.0.2
pyzmq==27.0.0
qdldl==0.1.7.post1
qpsolvers==4.8.1
qrcode==7.4.2
quadprog==0.1.13
ray==2.51.0
referencing==0.37.0
regex==2025.10.23
requests==2.32.3
requests-oauthlib==1.3.1
rerun-sdk==0.20.1
rich==14.2.0
rich-click==1.9.4
rl_games @ git+https://github.com/isaac-sim/rl_games.git@6b3534f29568158e9e29ec8bf83cc88fce5f0cae
robomimic @ git+https://github.com/ARISE-Initiative/robomimic.git@7c66e7a41b5d9dcc905b1a68346bfee1b49b79c9
rpds-py==0.28.0
rsl-rl-lib==2.3.3
Rtree==1.3.0
s3transfer==0.11.0
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.15.3
selenium==4.14.0
sentry-sdk==2.42.1
setproctitle==1.3.7
six==1.17.0
skrl==1.4.3
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
sshkeyboard==2.3.1
stable_baselines3==2.7.0
stack-data==0.6.3
starlette==0.45.3
sympy==1.13.3
-e git+https://github.com/silencht/televuer@86367f8c9ba248f4065b959bfc53e5f36339bf6d#egg=televuer&subdirectory=../../../teleop/televuer
tenacity==9.1.2
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.4
termcolor==3.2.0
threadpoolctl==3.6.0
tokenizers==0.22.1
toml==0.10.2
tomli==2.3.0
torch==2.7.0
torchaudio==2.7.0
torchvision==0.22.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.57.1
trimesh==4.5.1
trio==0.31.0
trio-websocket==0.12.2
triton==3.3.0
typing-inspection==0.4.2
typing_extensions==4.12.2
tzdata==2025.2
u-msgpack-python==2.8.0
-e git+https://github.com/unitreerobotics/unitree_sdk2_python@18b9ef8e57e69b73c5e9b301b8481fce16a44b9b#egg=unitree_sdk2py
urllib3==2.5.0
uvicorn==0.29.0
vuer==0.0.60
wandb==0.22.3
warp-lang==1.9.1
watchdog==4.0.0
waterbear==2.6.8
wcwidth==0.2.14
webbot==0.34
websockets==12.0
Werkzeug==3.1.3
widgetsnbextension==4.0.14
wrapt==1.16.0
wsproto==1.2.0
yarl==1.18.3
